requirements.txt
requests
python-dotenv
types-requests
